move sql query generation to datasource
